Implemented symbol tables for each scope
The Listener class is basically the semantic actions for entering and
exiting certain blocks of code.  SymbolTableFactory is the task object
for creating symbol tables. Currently, symbol tables are being created
for each scope correctly but we need to add the actual variables as well
as the name of each function scope.