Now have IR code generating for LABEL and STORE
IRNode is a class for storing IR instructions. IRList is linked list for
storing IRNodes. IRListener is a seperate LITTLEListener implementation
(similar to the Listener for SymbolTables), however it is specifically
for generating IR code instead. The IRGenerator class is the Task for
generating the IRList. The TinyGenerator class is a currently non
implemented class for converting IR code to Tiny code.